removed custom hook and created a fucntion
